Move remaining_quantity on ads table
Think about #6 when pushing to prod

+++ database/migrations/2016_09_14_165630_create_stocks_table.php
@@ -13,13 +13,15 @@ class CreateStocksTable extends Migration
      */
     public function up()
     {
-        Schema::create('stocks', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('ad_id')->references('id')->on('ads');
-            $table->unsignedInteger('initial');
-            $table->unsignedInteger('remaining');
-            $table->timestamps();
-        });
+        if(!env('DB_CONNECTION') == 'sqlite') {
+            Schema::create('stocks', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('ad_id')->references('id')->on('ads');
+                $table->unsignedInteger('initial');
+                $table->unsignedInteger('remaining');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -29,6 +31,8 @@ class CreateStocksTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('stocks');
+        if(!env('DB_CONNECTION') == 'sqlite') {
+            Schema::dropIfExists('stocks');
+        }
     }
 }
